base classes created
TODO:
- Adjust namings
- Refactor other components
- Update test fixtures
