+++ COMETwebapp/COMETwebapp.csproj
@@ -40,4 +40,12 @@
       </Content>
     </ItemGroup>
 
+    <ItemGroup>
+      <UpToDateCheckInput Remove="Components\ReferenceData\SelectedReferenceDataItem.razor" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <_ContentIncludedByDefault Remove="Components\ReferenceData\SelectedReferenceDataItem.razor" />
+    </ItemGroup>
+
 </Project>
