+++ COMETwebapp/ViewModels/Components/ReferenceData/IMeasurementUnitsTableViewModel.cs
@@ -36,72 +36,16 @@ namespace COMETwebapp.ViewModels.Components.ReferenceData
     /// <summary>
     /// View model used to manage <see cref="MeasurementUnit" />s
     /// </summary>
-    public interface IMeasurementUnitsTableViewModel : IApplicationBaseViewModel, IHaveReusableRows
+    public interface IMeasurementUnitsTableViewModel : IReferenceDataItemViewModel<MeasurementUnit>
     {
-        /// <summary>
-        /// Gets or sets the data source for the grid control.
-        /// </summary>
-        SourceList<MeasurementUnit> DataSource { get; }
-
-        /// <summary>
-        /// A reactive collection of <see cref="MeasurementUnitRowViewModel" />
-        /// </summary>
-        SourceList<MeasurementUnitRowViewModel> Rows { get; }
-
-        /// <summary>
-        /// Injected property to get access to <see cref="IShowHideDeprecatedThingsService" />
-        /// </summary>
-        IShowHideDeprecatedThingsService ShowHideDeprecatedThingsService { get; }
-
-        /// <summary>
-        /// The <see cref="MeasurementUnit" /> to create or edit
-        /// </summary>
-        MeasurementUnit MeasurementUnit { get; set; }
-
-        /// <summary>
-        /// Indicates if confirmation popup is visible
-        /// </summary>
-        bool IsOnDeprecationMode { get; set; }
-
-        /// <summary>
-        /// popup message dialog
-        /// </summary>
-        string PopupDialog { get; set; }
-
         /// <summary>
         /// Available <see cref="ReferenceDataLibrary" />s
         /// </summary>
         IEnumerable<ReferenceDataLibrary> ReferenceDataLibraries { get; set; }
 
         /// <summary>
-        /// Method invoked when confirming the deprecation/un-deprecation of a <see cref="MeasurementUnitsTableViewModel.MeasurementUnit" />
+        /// A reactive collection of <see cref="MeasurementUnitRowViewModel" />
         /// </summary>
-        /// <returns>A <see cref="Task" /></returns>
-        Task OnConfirmPopupButtonClick();
-
-        /// <summary>
-        /// Method invoked when canceling the deprecation/un-deprecation of a <see cref="MeasurementUnitsTableViewModel.MeasurementUnit" />
-        /// </summary>
-        void OnCancelPopupButtonClick();
-
-        /// <summary>
-        /// Action invoked when the deprecate or undeprecate button is clicked
-        /// </summary>
-        /// <param name="measurementUnitRow"> The <see cref="MeasurementUnitRowViewModel" /> to deprecate or undeprecate </param>
-        void OnDeprecateUnDeprecateButtonClick(MeasurementUnitRowViewModel measurementUnitRow);
-
-        /// <summary>
-        /// Method invoked when the component is ready to start, having received its
-        /// initial parameters from its parent in the render tree.
-        /// Override this method if you will perform an asynchronous operation and
-        /// want the component to refresh when that operation is completed.
-        /// </summary>
-        void InitializeViewModel();
-
-        /// <summary>
-        /// Tries to deprecate or undeprecate a <see cref="MeasurementUnitsTableViewModel.MeasurementUnit" />
-        /// </summary>
-        /// <returns>A <see cref="Task" /></returns>
-        Task DeprecatingOrUnDeprecatingMeasurementUnit();
+        SourceList<MeasurementUnitRowViewModel> Rows { get; }
     }
 }
