+++ COMETwebapp/ViewModels/Components/ReferenceData/MeasurementUnitsTableViewModel.cs
@@ -46,38 +46,8 @@ namespace COMETwebapp.ViewModels.Components.ReferenceData
     /// <summary>
     /// View model used to manage <see cref="MeasurementUnit" />s
     /// </summary>
-    public class MeasurementUnitsTableViewModel : ApplicationBaseViewModel, IMeasurementUnitsTableViewModel
+    public class MeasurementUnitsTableViewModel : ReferenceDataItemViewModel<MeasurementUnit>, IMeasurementUnitsTableViewModel
     {
-        /// <summary>
-        /// Injected property to get access to <see cref="IPermissionService" />
-        /// </summary>
-        private readonly IPermissionService permissionService;
-
-        /// <summary>
-        /// Injected property to get access to <see cref="ISessionService" />
-        /// </summary>
-        private readonly ISessionService sessionService;
-
-        /// <summary>
-        /// Backing field for <see cref="IsOnDeprecationMode" />
-        /// </summary>
-        private bool isOnDeprecationMode;
-
-        /// <summary>
-        /// The <see cref="ILogger{TCategoryName}"/>
-        /// </summary>
-        private readonly ILogger<MeasurementUnitsTableViewModel> logger;
-
-        /// <summary>
-        /// A collection of <see cref="Type" /> used to create <see cref="ObjectChangedEvent" /> subscriptions
-        /// </summary>
-        private static readonly IEnumerable<Type> ObjectChangedTypesOfInterest = new List<Type>
-        {
-            typeof(MeasurementUnit),
-            typeof(PersonRole),
-            typeof(ReferenceDataLibrary)
-        };
-
         /// <summary>
         /// Initializes a new instance of the <see cref="MeasurementUnitsTableViewModel" /> class.
         /// </summary>
@@ -86,21 +56,9 @@ namespace COMETwebapp.ViewModels.Components.ReferenceData
         /// <param name="messageBus">The <see cref="ICDPMessageBus"/></param>
         /// <param name="logger">The <see cref="ILogger{TCategoryName}"/></param>
         public MeasurementUnitsTableViewModel(ISessionService sessionService, IShowHideDeprecatedThingsService showHideDeprecatedThingsService, ICDPMessageBus messageBus, 
-            ILogger<MeasurementUnitsTableViewModel> logger) : base(sessionService, messageBus)
+            ILogger<MeasurementUnitsTableViewModel> logger) : base(sessionService, messageBus, showHideDeprecatedThingsService, logger)
         {
-            this.sessionService = sessionService;
-            this.permissionService = sessionService.Session.PermissionService;
-            this.ShowHideDeprecatedThingsService = showHideDeprecatedThingsService;
-            this.logger = logger;
-
-            this.InitializeSubscriptions(ObjectChangedTypesOfInterest);
-            this.RegisterViewModelWithReusableRows(this);
         }
-
-        /// <summary>
-        /// The <see cref="MeasurementUnit" /> to create or edit
-        /// </summary>
-        public MeasurementUnit MeasurementUnit { get; set; } = new SimpleUnit();
 
         /// <summary>
         /// Available <see cref="ReferenceDataLibrary" />s
@@ -108,205 +66,27 @@ namespace COMETwebapp.ViewModels.Components.ReferenceData
         public IEnumerable<ReferenceDataLibrary> ReferenceDataLibraries { get; set; }
 
         /// <summary>
-        /// Injected property to get access to <see cref="IShowHideDeprecatedThingsService" />
-        /// </summary>
-        public IShowHideDeprecatedThingsService ShowHideDeprecatedThingsService { get; }
-
-        /// <summary>
         /// A reactive collection of <see cref="MeasurementUnitRowViewModel" />
         /// </summary>
         public SourceList<MeasurementUnitRowViewModel> Rows { get; } = new();
 
-        /// <summary>
-        /// Gets or sets the data source for the grid control.
-        /// </summary>
-        public SourceList<MeasurementUnit> DataSource { get; } = new();
-
-        /// <summary>
-        /// Indicates if confirmation popup is visible
-        /// </summary>
-        public bool IsOnDeprecationMode
+        protected override void ConvertRowsToSpecificType()
         {
-            get => this.isOnDeprecationMode;
-            set => this.RaiseAndSetIfChanged(ref this.isOnDeprecationMode, value);
-        }
+            var convertedItems = this.internalRows.Items.Select(x => new MeasurementUnitRowViewModel(x.Thing)
+            {
+                IsAllowedToWrite = x.IsAllowedToWrite
+            });
 
-        /// <summary>
-        /// popup message dialog
-        /// </summary>
-        public string PopupDialog { get; set; }
+            this.Rows.EditDiff(convertedItems);
+        }
 
         /// <summary>
         /// Initializes the <see cref="MeasurementUnitsTableViewModel"/>
         /// </summary>
-        public void InitializeViewModel()
+        public override void InitializeViewModel()
         {
-            foreach (var referenceDataLibrary in this.sessionService.Session.RetrieveSiteDirectory().AvailableReferenceDataLibraries())
-            {
-                this.DataSource.AddRange(referenceDataLibrary.Unit);
-            }
-
-            this.ReferenceDataLibraries = this.sessionService.Session.RetrieveSiteDirectory().AvailableReferenceDataLibraries();
-            this.Rows.AddRange(this.DataSource.Items.Select(x => new MeasurementUnitRowViewModel(x)));
-            this.RefreshAccessRight();
-        }
-
-        /// <summary>
-        /// Method invoked when confirming the deprecation/un-deprecation of a <see cref="MeasurementUnit" />
-        /// </summary>
-        /// <returns>A <see cref="Task" /></returns>
-        public async Task OnConfirmPopupButtonClick()
-        {
-            await this.DeprecatingOrUnDeprecatingMeasurementUnit();
-            this.IsOnDeprecationMode = false;
-        }
-
-        /// <summary>
-        /// Method invoked when canceling the deprecation/un-deprecation of a <see cref="MeasurementUnit" />
-        /// </summary>
-        public void OnCancelPopupButtonClick()
-        {
-            this.IsOnDeprecationMode = false;
-        }
-
-        /// <summary>
-        /// Action invoked when the deprecate or undeprecate button is clicked
-        /// </summary>
-        /// <param name="measurementUnitRow"> The <see cref="MeasurementUnitRowViewModel" /> to deprecate or undeprecate </param>
-        public void OnDeprecateUnDeprecateButtonClick(MeasurementUnitRowViewModel measurementUnitRow)
-        {
-            this.MeasurementUnit = measurementUnitRow.MeasurementUnit;
-            this.PopupDialog = this.MeasurementUnit.IsDeprecated ? "You are about to un-deprecate the measurement unit: " + measurementUnitRow.Name : "You are about to deprecate the measurement unit: " + measurementUnitRow.Name;
-            this.IsOnDeprecationMode = true;
-        }
-
-        /// <summary>
-        /// Tries to deprecate or undeprecate a <see cref="MeasurementUnit" />
-        /// </summary>
-        /// <returns>A <see cref="Task" /></returns>
-        public async Task DeprecatingOrUnDeprecatingMeasurementUnit()
-        {
-            var siteDirectoryClone = this.sessionService.GetSiteDirectory().Clone(false);
-            var clonedMeasurementUnit = this.MeasurementUnit.Clone(false);
-            clonedMeasurementUnit.IsDeprecated = !clonedMeasurementUnit.IsDeprecated;
-
-            try
-            {
-                await this.sessionService.UpdateThings(siteDirectoryClone, clonedMeasurementUnit);
-            }
-            catch (Exception exception)
-            {
-                this.logger.LogError(exception, "An error has occurred while trying to deprecating or un-deprecating the measurement unit {unitName}", clonedMeasurementUnit.ShortName);
-            }
-        }
-
-        /// <summary>
-        /// Add rows related to <see cref="Thing" /> that has been added
-        /// </summary>
-        /// <param name="addedThings">A collection of added <see cref="Thing" /></param>
-        public void AddRows(IEnumerable<Thing> addedThings)
-        {
-            var addedMeasurementUnits = addedThings.OfType<MeasurementUnit>().ToList();
-            this.Rows.AddRange(addedMeasurementUnits.Select(x => new MeasurementUnitRowViewModel(x)));
-        }
-
-        /// <summary>
-        /// Updates rows related to <see cref="Thing" /> that have been updated
-        /// </summary>
-        /// <param name="updatedThings">A collection of updated <see cref="Thing" /></param>
-        public void UpdateRows(IEnumerable<Thing> updatedThings)
-        {
-            var updatedThingsList = updatedThings.ToList();
-
-            var updatedMeasurementUnits = updatedThingsList.OfType<MeasurementUnit>();
-            var updatedPersonRoles = updatedThingsList.OfType<PersonRole>();
-            var updatedRdls = updatedThingsList.OfType<ReferenceDataLibrary>();
-
-            this.Rows.Edit(action =>
-            {
-                foreach (var updatedMeasurementUnit in updatedMeasurementUnits)
-                {
-                    var updatedRow = new MeasurementUnitRowViewModel(updatedMeasurementUnit);
-                    var rowToUpdate = this.Rows.Items.First(x => x.MeasurementUnit.Iid == updatedMeasurementUnit.Iid);
-                    action.Replace(rowToUpdate, updatedRow);
-                }
-            });
-
-            foreach (var rdl in updatedRdls)
-            {
-                this.RefreshContainerName(rdl);
-            }
-
-            if (updatedPersonRoles.Any())
-            {
-                this.RefreshAccessRight();
-            }
-        }
-
-        /// <summary>
-        /// Remove rows related to a <see cref="Thing" /> that has been deleted
-        /// </summary>
-        /// <param name="deletedThings">A collection of deleted <see cref="Thing" /></param>
-        public void RemoveRows(IEnumerable<Thing> deletedThings)
-        {
-            var measurementUnitsIidsToRemove = deletedThings.OfType<MeasurementUnit>().Select(x => x.Iid);
-            var rowsToDelete = this.Rows.Items.Where(x => measurementUnitsIidsToRemove.Contains(x.MeasurementUnit.Iid)).ToList();
-
-            this.Rows.RemoveMany(rowsToDelete);
-        }
-
-        /// <summary>
-        /// Handles the refresh of the current <see cref="ISession" />
-        /// </summary>
-        /// <returns>A <see cref="Task" /></returns>
-        protected override Task OnSessionRefreshed()
-        {
-            if (this.AddedThings.Count == 0 && this.UpdatedThings.Count == 0 && this.DeletedThings.Count == 0)
-            {
-                return Task.CompletedTask;
-            }
-
-            this.IsLoading = true;
-            this.UpdateInnerComponents();
-            this.RefreshAccessRight();
-            this.ClearRecordedChanges();
-            this.IsLoading = false;
-
-            return Task.CompletedTask;
-        }
-
-        /// <summary>
-        /// Refresh the displayed container name for the measurement unit rows
-        /// </summary>
-        /// <param name="rdl">
-        /// The updated <see cref="ReferenceDataLibrary" />.
-        /// </param>
-        private void RefreshContainerName(ReferenceDataLibrary rdl)
-        {
-            this.IsLoading = true;
-            var rowsContainedByUpdatedRdl = this.Rows.Items.Where(x => x.MeasurementUnit.Container.Iid == rdl.Iid);
-
-            foreach (var measurementUnit in rowsContainedByUpdatedRdl)
-            {
-                measurementUnit.ContainerName = rdl.ShortName;
-            }
-
-            this.IsLoading = false;
-        }
-
-        /// <summary>
-        /// Updates the active user access rights
-        /// </summary>
-        private void RefreshAccessRight()
-        {
-            this.IsLoading = true;
-
-            foreach (var row in this.Rows.Items)
-            {
-                row.IsAllowedToWrite = this.permissionService.CanWrite(ClassKind.MeasurementUnit, row.MeasurementUnit.Container);
-            }
-
-            this.IsLoading = false;
+            base.InitializeViewModel();
+            this.ReferenceDataLibraries = this.SessionService.Session.RetrieveSiteDirectory().AvailableReferenceDataLibraries();
         }
     }
 }
