+++ COMETwebapp/appsettings.json
@@ -2,7 +2,7 @@
   "AllowedHosts": "*",
   "StringTablePath": "wwwroot/DefaultTextConfiguration.json",
   "ServerConfiguration": {
-    "ServerAddress": "",
+    "ServerAddress": "http://localhost:5000/",
     "BookInputConfiguration": {
       "ShowName": true,
       "ShowShortName": true
